pricing int to float

+++ resources/views/all_products.blade.php
@@ -417,7 +417,7 @@
                             var price = parseFloat(item.price);
                             var quantity = parseFloat(item.quantity);
 
-                            var subtotal = parseInt(price * quantity);
+                            var subtotal = parseFloat(price * quantity);
                             var cart_total = cart_total + subtotal;
                             var total_cart_quantity = total_cart_quantity + quantity;
                             $('#subtotal_' + product_id).html('$'+subtotal);
